refactor: 패키지 구조 분리 및 액티비티 명 수정
