+++ app/src/main/AndroidManifest.xml
@@ -12,13 +12,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".MovieTicketActivity"
+            android:name=".activity.TicketResultActivity"
             android:exported="false" />
         <activity
-            android:name=".TicketingActivity"
+            android:name=".activity.MovieDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".MovieListActivity"
+            android:name=".activity.MovieListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
