+++ app/src/main/java/woowacourse/movie/model/PlayingTimes.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.model
 
 import java.time.DayOfWeek
 import java.time.LocalDate
