+++ app/src/main/java/woowacourse/movie/model/policy/MorningPolicy.kt
@@ -1,14 +1,15 @@
-package woowacourse.movie.domain
+package woowacourse.movie.model.policy
 
+import woowacourse.movie.model.Price
 import java.time.LocalDate
 import java.time.LocalTime
 
-class NightPolicy : DiscountPolicy() {
+class MorningPolicy : DiscountPolicy() {
     override fun getDiscountPrice(price: Price): Price {
         return Price(price.price - 2000)
     }
 
     override fun isAvailable(date: LocalDate, time: LocalTime): Boolean {
-        return time.hour >= 20
+        return time.hour < 11
     }
 }
