+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".TicketingActivity">
+    tools:context=".activity.MovieDetailActivity">
 
     <ScrollView
         android:layout_width="match_parent"
