+++ app/src/main/res/layout/activity_ticket_result.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".MovieTicketActivity">
+    tools:context=".activity.TicketResultActivity">
 
     <TextView
         android:layout_width="match_parent"
