+++ app/src/test/java/woowacourse/movie/PlayingTimesTest.kt
@@ -2,6 +2,7 @@ package woowacourse.movie
 
 import junit.framework.TestCase.assertEquals
 import org.junit.Test
+import woowacourse.movie.model.PlayingTimes
 import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.LocalTime
